+++ helm_deploy/hmpps-book-a-video-link-ui/values.yaml
@@ -63,6 +63,7 @@ generic-service:
     feature-toggles:
       FEATURE_ENHANCED_PROBATION_JOURNEY: "FEATURE_MASTER_VLPM_TYPES"
       FEATURE_ADMIN_LOCATION_DECORATION: "FEATURE_ADMIN_LOCATION_DECORATION"
+      FEATURE_ALTERED_COURT_JOURNEY: "FEATURE_ALTERED_COURT_JOURNEY"
 
   allowlist:
     groups:
