+++ server/routes/journeys/bookAVideoLink/court/createRoutes.ts
@@ -4,14 +4,14 @@ import type { Services } from '../../../../services'
 import { PageHandler } from '../../../interfaces/pageHandler'
 import logPageViewMiddleware from '../../../../middleware/logPageViewMiddleware'
 import validationMiddleware from '../../../../middleware/validationMiddleware'
-import NewBookingHandler from './handlers/newBookingHandler'
+import DeprecatedNewBookingHandler from './handlers/deprecatedNewBookingHandler'
 import CheckBookingHandler from './handlers/checkBookingHandler'
 import ConfirmationHandler from './handlers/confirmationHandler'
-import BookingNotAvailableHandler from './handlers/bookingNotAvailableHandler'
+import DeprecatedBookingNotAvailableHandler from './handlers/deprecatedBookingNotAvailableHandler'
 import config from '../../../../config'
-import V2NewBookingHandler from './handlers/v2NewBookingHandler'
-import V2SelectRoomsHandler from './handlers/v2SelectRoomsHandler'
-import V2NoRoomsHandler from './handlers/v2NoRoomsHandler'
+import BookingDetailsHandler from './handlers/bookingDetailsHandler'
+import SelectRoomsHandler from './handlers/selectRoomsHandler'
+import BookingNotAvailableHandler from './handlers/bookingNotAvailableHandler'
 
 export default function CreateRoutes({
   auditService,
@@ -30,24 +30,20 @@ export default function CreateRoutes({
     router.post(path, validationMiddleware(handler.BODY), asyncMiddleware(handler.POST))
 
   if (config.featureToggles.alteredCourtJourneyEnabled) {
-    /**
-     * The 3-stage journey to enter date and time and which hearings are required, and then the 2nd stage
-     * to select from lists of available rooms returned from the API DateTimeAvailability check. If any
-     * room list is empty, a 3rd stage presents a no-rooms-available form to prompt for a change of date and time.
-     */
     route(
       `${basePath}/video-link-booking`,
-      new V2NewBookingHandler(courtsService, prisonerService, referenceDataService),
+      new BookingDetailsHandler(courtsService, prisonerService, referenceDataService),
     )
-    route(
-      `${basePath}/video-link-booking/select-rooms`,
-      new V2SelectRoomsHandler(courtsService, courtBookingService, prisonerService),
-    )
-    route(`${basePath}/video-link-booking/no-rooms`, new V2NoRoomsHandler(courtsService, prisonerService))
   } else {
     route(
       `${basePath}/video-link-booking`,
-      new NewBookingHandler(courtsService, prisonService, prisonerService, referenceDataService, videoLinkService),
+      new DeprecatedNewBookingHandler(
+        courtsService,
+        prisonService,
+        prisonerService,
+        referenceDataService,
+        videoLinkService,
+      ),
     )
   }
 
@@ -62,18 +58,23 @@ export default function CreateRoutes({
     return next()
   })
 
+  if (config.featureToggles.alteredCourtJourneyEnabled) {
+    route(
+      `${basePath}/video-link-booking/select-rooms`,
+      new SelectRoomsHandler(courtsService, courtBookingService, prisonerService),
+    )
+    route(
+      `${basePath}/video-link-booking/not-available`,
+      new BookingNotAvailableHandler(courtsService, prisonerService),
+    )
+  } else {
+    route(`${basePath}/video-link-booking/not-available`, new DeprecatedBookingNotAvailableHandler(courtBookingService))
+  }
+
   route(
     `${basePath}/video-link-booking/check-booking`,
     new CheckBookingHandler(courtBookingService, courtsService, prisonService, referenceDataService, videoLinkService),
   )
-
-  /**
-   * The old availability check is offered if no bookings are available when the feature switch is off as
-   * this does not take room decoration into account.
-   */
-  if (!config.featureToggles.alteredCourtJourneyEnabled) {
-    route(`${basePath}/video-link-booking/not-available`, new BookingNotAvailableHandler(courtBookingService))
-  }
 
   return router
 }
