+++ server/routes/journeys/bookAVideoLink/court/requestRoutes.ts
@@ -5,9 +5,9 @@ import { PageHandler } from '../../../interfaces/pageHandler'
 import logPageViewMiddleware from '../../../../middleware/logPageViewMiddleware'
 import validationMiddleware from '../../../../middleware/validationMiddleware'
 import PrisonerDetailsHandler from './handlers/prisonerDetailsHandler'
-import NewBookingHandler from './handlers/newBookingHandler'
+import DeprecatedNewBookingHandler from './handlers/deprecatedNewBookingHandler'
 import CheckBookingHandler from './handlers/checkBookingHandler'
-import BookingNotAvailableHandler from './handlers/bookingNotAvailableHandler'
+import DeprecatedBookingNotAvailableHandler from './handlers/deprecatedBookingNotAvailableHandler'
 import BookingRequestedHandler from './handlers/bookingRequestedHandler'
 
 export default function RequestRoutes({
@@ -36,13 +36,19 @@ export default function RequestRoutes({
 
   route(
     '/prisoner/video-link-booking',
-    new NewBookingHandler(courtsService, prisonService, prisonerService, referenceDataService, videoLinkService),
+    new DeprecatedNewBookingHandler(
+      courtsService,
+      prisonService,
+      prisonerService,
+      referenceDataService,
+      videoLinkService,
+    ),
   )
   route(
     `/prisoner/video-link-booking/check-booking`,
     new CheckBookingHandler(courtBookingService, courtsService, prisonService, referenceDataService, videoLinkService),
   )
-  route(`/prisoner/video-link-booking/not-available`, new BookingNotAvailableHandler(courtBookingService))
+  route(`/prisoner/video-link-booking/not-available`, new DeprecatedBookingNotAvailableHandler(courtBookingService))
 
   return router
 }
