+++ server/services/auditService.ts
@@ -10,7 +10,6 @@ export enum Page {
   PRISONER_SEARCH_RESULTS_PAGE = 'PRISONER_SEARCH_RESULTS_PAGE',
   BOOKING_DETAILS_PAGE = 'BOOKING_DETAILS_PAGE',
   SELECT_ROOMS_PAGE = 'SELECT_ROOMS_PAGE',
-  ALTERNATIVE_ROOMS_PAGE = 'ALTERNATIVE_ROOMS_PAGE',
   CHECK_BOOKING_PAGE = 'CHECK_BOOKING_PAGE',
   COMMENTS_PAGE = 'COMMENTS_PAGE',
   BOOKING_AVAILABILITY_PAGE = 'BOOKING_AVAILABILITY_PAGE',
