+++ server/views/pages/bookAVideoLink/court/notAvailableDeprecated.njk
@@ -1,0 +1,80 @@
+{% extends "partials/layout.njk" %}
+
+{% from "govuk/components/details/macro.njk" import govukDetails %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
+
+{% set pageTitle = "Video link booking not available" %}
+
+{% block content %}
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-three-quarters">
+            <h1 class="govuk-heading-l">{{ pageTitle }}</h1>
+        </div>
+    </div>
+
+    <div class="govuk-grid-row">
+        <div class="govuk-grid-column-three-quarters">
+            <div class="govuk-body">
+                <p>There are no video link bookings available at that time in the rooms selected.</p>
+                <p>You can choose a different date or time or search using different rooms allocated to your court.</p>
+            </div>
+
+            {{ govukDetails({
+                summaryText: "What to do if you can't find a booking",
+                text: "You should contact the prison you're trying to book a room in. They may be able to find an alternative room for your booking."
+            }) }}
+
+            {{ govukButton({
+                text: "Search again",
+                href: "../video-link-booking"
+            }) }}
+        </div>
+    </div>
+
+    {% if alternatives.length > 0 %}
+        <hr class="govuk-section-break govuk-section-break--m govuk-section-break--visible">
+        <div class="govuk-grid-row">
+            <div class="govuk-grid-column-two-thirds">
+                <h2 class="govuk-heading-m">Other available bookings in selected rooms</h2>
+                {% for alternative in alternatives %}
+                    {% set rows = [
+                        {
+                            key:   { text: "Date" },
+                            value: { text: session.journey.bookACourtHearing.date | formatDate }
+                        },
+                        {
+                            key:   { text: "Court hearing start time" },
+                            value: { text: alternative.main.interval.start }
+                        },
+                        {
+                            key:   { text: "Court hearing end time" },
+                            value: { text: alternative.main.interval.end }
+                        },
+                        {
+                            key:   { text: "Pre-court hearing briefing" },
+                            value: { text: alternative.pre.interval.start + " to " + alternative.pre.interval.end }
+                        } if alternative.pre,
+                        {
+                            key:   { text: "Post-court hearing briefing" },
+                            value: { text: alternative.post.interval.start + " to " + alternative.post.interval.end }
+                        } if alternative.post
+                    ] %}
+
+                    {{ govukSummaryList({ rows: rows }) }}
+
+                    <form method="POST">
+                        <input type="hidden" name="_csrf" value="{{ csrfToken }}" />
+                        <input type="hidden" name="startTime" value="{{ alternative.main.interval.start }}" />
+                        <input type="hidden" name="endTime" value="{{ alternative.main.interval.end }}" />
+                        <input type="hidden" name="preStart" value="{{ alternative.pre.interval.start }}" />
+                        <input type="hidden" name="preEnd" value="{{ alternative.pre.interval.end }}" />
+                        <input type="hidden" name="postStart" value="{{ alternative.post.interval.start }}" />
+                        <input type="hidden" name="postEnd" value="{{ alternative.post.interval.end }}" />
+                        {{ govukButton({ text: "Book video link", type: "submit", classes: 'govuk-button govuk-button--secondary', attributes: {'data-qa': 'option-' + loop.index} }) }}
+                    </form>
+                {% endfor %}
+            </div>
+        </div>
+    {% endif %}
+{% endblock %}
