--- server/views/pages/bookAVideoLink/court/v2NoRooms.njk
@@ -1,124 +1,0 @@
-{% extends "partials/layout.njk" %}
-
-{% from "components/hmppsTimePicker.njk" import hmppsTimePicker %}
-{% from "moj/components/date-picker/macro.njk" import mojDatePicker %}
-{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
-{% from "govuk/components/select/macro.njk" import govukSelect %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk/components/radios/macro.njk" import govukRadios %}
-{% from "govuk/components/input/macro.njk" import govukInput %}
-
-{% set mode = session.req.params.mode %}
-{% set pageTitle = 'No bookings available for your selected time' %}
-
-{% block content %}
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-three-quarters">
-            <h1 class="govuk-heading-l">{{ pageTitle }}</h1>
-            <p>
-                There may be available bookings at other times on the selected date. Please change the date or time.
-            </p>
-        </div>
-
-    </div>
-
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-three-quarters">
-            <form method='POST'>
-                <input type="hidden" name="_csrf" value="{{ csrfToken }}" />
-
-                {{ govukSummaryList({
-                    classes: "govuk-summary-list--no-border",
-                    rows: [
-                        {
-                            key: {
-                            text: "Name"
-                        },
-                            value: {
-                            text: ((prisoner.firstName + " " + prisoner.lastName) | convertToTitleCase) + ((' (' + prisoner.prisonerNumber + ')') if prisoner.prisonerNumber)
-                        }
-                        },
-                        {
-                            key: {
-                            text: "Prison"
-                        },
-                            value: {
-                            text: prisoner.prisonName
-                        }
-                        },
-                        {
-                            key: {
-                            text: "Court"
-                        },
-                            value: {
-                            text: courts[0].description
-                        }
-                        } if courts.length == 1,
-                        {
-                            key: {
-                            text: "Court"
-                        },
-                            value: {
-                            text: (courts | find('code', session.journey.bookACourtHearing.courtCode)).description
-                        }
-                        } if courts.length > 1,
-                        {
-                            key: {
-                            text: "Date"
-                        },
-                            value: {
-                            text: session.journey.bookACourtHearing.date | formatDate('dd/MM/yyyy')
-                        }
-                        },
-                        {
-                            key: {
-                            text: "Hearing time"
-                        },
-                            value: {
-                            text: mainTimes
-                        }
-                        },
-                        {
-                            key: {
-                            text: "Pre-court hearing briefing time"
-                        },
-                            value: {
-                            text: preTimes
-                        }
-                        } if preTimes,
-                        {
-                            key: {
-                            text: "Post-court hearing briefing time"
-                        },
-                            value: {
-                            text: postTimes
-                        }
-                        } if postTimes
-                    ]
-                }) }}
-
-                <div class="govuk-button-group">
-                    {{ govukButton({
-                        text: "Change times",
-                        preventDoubleClick: true,
-                        type: "submit"
-                    }) }}
-
-                    {{ govukButton({
-                        text: "Cancel",
-                        classes: "govuk-button--secondary",
-                        preventDoubleClick: true,
-                        href: "/"
-                    }) }}
-                </div>
-
-            </form>
-
-            {% if mode == 'create' or mode == 'request' %}
-                <a href="../../../../prisoner-search/search" class="govuk-link govuk-link--no-visited-state govuk-!-font-size-19">Return to prisoner search</a>
-            {% else %}
-                <a href="/court/view-booking/{{ session.journey.bookACourtHearing.bookingId }}" class="govuk-link govuk-link--no-visited-state govuk-!-font-size-19">Cancel</a>
-            {% endif %}
-        </div>
-    </div>
-{% endblock %}
